requirements.txt
apturl==0.5.2
argostranslate==1.9.6
asgiref==3.5.0
attrs==23.2.0
Automat==22.10.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
ctranslate2==4.2.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.3
fsspec==2024.6.0
future==0.18.2
grpcio==1.64.1
grpcio-tools==1.64.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==24.0
paramiko==2.9.3
parsel==1.9.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
Protego==0.3.1
protobuf==5.27.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyDispatcher==2.0.7
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
queuelib==1.7.0
regex==2024.5.15
reportlab==3.6.8
requests==2.25.1
requests-file==2.0.0
sacremoses==0.0.53
Scrapy==2.11.2
SecretStorage==3.3.1
sentencepiece==0.2.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.2
srt==3.5.3
stanza==1.1.1
sympy==1.12.1
systemd-python==234
terminator==2.1.1
tldextract==5.1.2
torch==2.3.1
tqdm==4.66.4
triton==2.3.1
Twisted==24.3.0
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
vosk==0.3.45
w3lib==2.1.2
wadllib==1.3.6
websockets==12.0
Werkzeug==3.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.4
